requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
arrow==1.3.0
    # via isoduration
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
datasynthtool @ git+https://github.com/slott56/DataSynthTool.git@v2024.8
    # via -r requirements.in
fqdn==1.5.1
    # via jsonschema
functional-sql @ git+https://github.com/slott56/functional-SQL.git@v2024.8
    # via -r requirements.in
idna==3.7
    # via jsonschema
isoduration==20.11.0
    # via jsonschema
jsonpointer==3.0.0
    # via jsonschema
jsonschema[format]==4.23.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
pydantic==2.8.2
    # via -r requirements.in
pydantic-core==2.20.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   arrow
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rfc3339-validator==0.1.4
    # via jsonschema
rfc3987==1.3.8
    # via jsonschema
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   python-dateutil
    #   rfc3339-validator
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
uri-template==1.3.0
    # via jsonschema
webcolors==24.6.0
    # via jsonschema
